Moved file tests to corresponding package
